+++ test/TestMutableValue.php
@@ -18,7 +18,7 @@
 
 namespace CloudCreativity\Utils\Value;
 
-final class TestValue extends AbstractValue
+final class TestMutableValue extends AbstractMutableValue
 {
 
     /**
@@ -26,14 +26,6 @@ final class TestValue extends AbstractValue
      */
     protected function accept($value)
     {
-        return is_string($value) && 2 < strlen($value);
-    }
-
-    /**
-     * @inheritdoc
-     */
-    protected function useStrict()
-    {
-        return false;
+        return is_int($value);
     }
 }