+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
-## [0.1.0] - 2016/05/14
+## [0.2.0] - 2016-10-12
+
+### Changed
+- Now only supports scalar value objects, array value interfaces have been removed.
+- Scalar values are now non-mutable by default, with mutability denoted by a specific interface.
+
+## [0.1.0] - 2016-05-14
 Initial release.
 
 ### Added